+++ webviz_subsurface/_utils/webvizstore_functions.py
@@ -1,0 +1,19 @@
+from pathlib import Path
+import io
+import json
+
+from webviz_config.webviz_store import webvizstore
+
+
+@webvizstore
+def get_path(path: str) -> Path:
+    return Path(path)
+
+
+@webvizstore
+def find_files(folder: Path, suffix: str) -> io.BytesIO:
+    return io.BytesIO(
+        json.dumps(
+            sorted([str(filename) for filename in folder.glob(f"*{suffix}")])
+        ).encode()
+    )